# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tournament Management API",
        "description": "Tournament Management API for Android App. This API provides endpoints for managing tournaments, user authentication, and player interactions.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/player/dashboard": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Get player dashboard data",
                "description": "REQUIRES: Authorization header with Bearer token. Format: 'Bearer {token}'. Example: 'Bearer 6|YuxRBo5pyMcpMOraGyRKfNZnqKsVuwQ7dG5jKu4Dab7b101f'",
                "operationId": "0119bdffa2df355020c8e406823d0d60",
                "responses": {
                    "200": {
                        "description": "Dashboard data"
                    },
                    "401": {
                        "description": "Unauthenticated - Missing or invalid Bearer token"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/player/profile": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Get player profile",
                "operationId": "80c5238026cbf9eae6449cb01cbcf89f",
                "responses": {
                    "200": {
                        "description": "Player profile"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Player"
                ],
                "summary": "Update player profile",
                "operationId": "f5565bea0358dd7d9929b998e65768c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/player/tournaments": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "List published tournaments",
                "operationId": "a500cf3945fb3bd165fbae6435521eda",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by sports category",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of published tournaments"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/player/tournaments/{id}": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Get tournament basic details (location hidden)",
                "operationId": "ba17d3c342c7d678b0a33020c4321e3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tournament basic details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/player/tournaments/{id}/interest": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Express interest in tournament",
                "operationId": "f98483c6ea386a7d1f4bd352413ebdc1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Interest expressed successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/player/tournaments/{id}/subscribe": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Subscribe to tournament",
                "operationId": "6a38dd99de8220ddc90ab2d3a6c8decd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Subscription created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/player/tournaments/{id}/details": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Get full tournament details (requires subscription)",
                "operationId": "9f1e69eb64a4b25d28e886997b067106",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Full tournament details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/player/subscriptions/{id}/pay": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Pay for subscription",
                "operationId": "41d12c714249117464863dbac4938557",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_method": {
                                        "type": "string",
                                        "example": "card"
                                    },
                                    "payment_details": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment processed successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "REQUIRED: Bearer token authentication. Format: \"Bearer {token}\". Example: \"Bearer 6|YuxRBo5pyMcpMOraGyRKfNZnqKsVuwQ7dG5jKu4Dab7b101f\". You MUST include the word \"Bearer\" followed by a space before your token.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}